fix: pin boundary jnp.eye fallback dtype to the Hessian's dtype
jnp.eye(dimensions) with no dtype= follows jax's ambient default
float dtype (governed by jax_enable_x64) rather than the dtype of
the hessian it stands in for at boundary points. When they diverge,
jnp.where(is_boundary, jnp.eye(...), hessian) mismatches, which the
minimize/maximize custom_jvp rules reject as a primal/tangent dtype
mismatch.

Pass dtype=hessian.dtype explicitly at all four call sites, matching
the existing jnp.eye(n, dtype=...) convention already used elsewhere
in these files (simplex_bernstein.py:292, rational_bernstein.py:434).

Fixes #4

+++ src/xbernstein/_tensor_bernstein.py
@@ -213,7 +213,9 @@ def _tensor_minimize_jvp(
         ]
     )
     is_boundary = jnp.any((point == 0.0) | (point == 1.0))
-    safe_hessian = jnp.where(is_boundary, jnp.eye(dimensions), hessian)
+    safe_hessian = jnp.where(
+        is_boundary, jnp.eye(dimensions, dtype=hessian.dtype), hessian
+    )
     safe_gradient_tangent = jnp.where(
         is_boundary, jnp.zeros_like(gradient_tangent), gradient_tangent
     )
